granule analysis macro
Suen et. al 2019 PIWI/DEPS-1 direct interaction is essential for piRNA function and condensate ultrastructure in Caenorhabditis elegans

This macro uses the HKM Segment plugin to map nuclei in channel 1 and granules in channels 2 and 3. It measures granule area, circularity, signal intensity, Pearson's R for the two granule channels, and distance from the nearest nucleus signal boundary (positive values are inside the signal area).
Difference of Gaussians (k=1.4 to approximate a Laplacian of Gaussian) pre-processing is applied before detecting granules. Functions for histogram and scatter plotting are included for convenience.